优化音频处理策略中的删除操作和波形保存方法
替换`os.remove`为`shutil.rmtree`以彻底删除目录及其中内容。调整波形文件保存方式，使用`soundfile`替代`torchaudio`进行保存，优化参数处理，并统一音频处理中的日志错误记录。这些更改旨在提升代码的稳健性和执行效率。

+++ uvr5/mdxnet.py
@@ -170,11 +170,14 @@ class Predictor:
                         -_ort.run(None, {"input": -spek.cpu().numpy()})[0] * 0.5
                         + _ort.run(None, {"input": spek.cpu().numpy()})[0] * 0.5
                     )
-                    tar_waves = model.istft(torch.tensor(spec_pred).to(cpu))
+                    tar_waves = model.istft(
+                        torch.tensor(spec_pred, dtype=torch.complex64).to(cpu)
+                    )
                 else:
                     tar_waves = model.istft(
                         torch.tensor(
-                            _ort.run(None, {"input": spek.cpu().numpy()})[0]
+                            _ort.run(None, {"input": spek.cpu().numpy()})[0],
+                            dtype=torch.complex64,
                         ).to(cpu)
                     )
                 tar_signal = (
